Implement a moving camera

+++ Box.h
@@ -18,7 +18,7 @@ class Box : public Primitive {
         virtual void intersect(HitRecord& hit, const RenderContext& context, const Ray& ray) const;
         virtual void normal(Vector& normal, const RenderContext& context,
                 const Point& hitpos, const Ray& ray, const HitRecord& hit) const;
-        virtual void preprocess(double maxTime);
+        virtual void preprocess(int maxTime);
     protected:
         Animation<Point> min;
         Animation<Point> max;
