+++ Object.h
@@ -13,7 +13,7 @@ class Object {
   Object();
   virtual ~Object();
 
-  virtual void preprocess(double maxTime);
+  virtual void preprocess(int maxTime);
   virtual void getBounds(BoundingBox& bbox, const RenderContext&) const = 0;
   virtual void intersect(HitRecord& hit, const RenderContext& context, const Ray& ray) const = 0;
