+++ Triangle.h
@@ -20,7 +20,7 @@ class Triangle : public Primitive {
         virtual void intersect(HitRecord& hit, const RenderContext& context, const Ray& ray) const;
         virtual void normal(Vector& normal, const RenderContext& context,
                 const Point& hitpos, const Ray& ray, const HitRecord& hit) const;
-        virtual void preprocess(double maxTime);
+        virtual void preprocess(int maxTime);
     protected:
         Animation<Point> p0, p1, p2;
         Animation<Vector> n;
